refactor(test): colocate tests with tested code
- Move test files from tests/ directory to src/ alongside the code they test
- Add testing module support infrastructure  
- Update imports to use crate:: paths
- Remove old test files
