+++ cat-launcher/src-tauri/src/testing/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+pub mod support;
